jira_zephyr_coverage_langgraph/requirements.txt
fastapi
uvicorn
httpx
openai
faiss-cpu
pydantic
pytest
